Add plausibilization callback
Some arguments' sanity cannot be determined at the time they're parsed.
For example, if option A or option B needs to be given, we only know
after we've parsed all of them if the given set was okay. This
introduces a callback that will be called in the end for all parameters
(this might be a dumb idea; maybe we'll change that in the future).